% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProject.R
\name{addProject}
\alias{addProject}
\title{addProject - A function for adding a project to the DB}
\usage{
addProject(dbName, projName, person = "Unknown")
}
\arguments{
\item{dbName}{The name of the .db file in which views for landmarks will be setup}

\item{projName}{The name of the project to be submitted to the DB}

\item{person}{The person who's in charge of this project}
}
\value{
Nothing
}
\description{
addProject - A function for adding a project to the DB
}
