% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDB.R
\name{buildDB}
\alias{buildDB}
\title{buildDB - A function for creating a new, empty .db file with the structure used}
\usage{
buildDB(dbName = "NameOfDB.db", path = getwd())
}
\arguments{
\item{dbName}{The desired name for the .db file which will be created}

\item{path}{Path to where DB is located, defaults to working directory}
}
\value{
Nothing
}
\description{
buildDB - A function for creating a new, empty .db file with the structure used
}
