% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOutliers.R
\name{calcOutliers}
\alias{calcOutliers}
\title{calcOutliers - A function for calculating outliers for checkLM}
\usage{
calcOutliers(Metric, alpha = 0.01)
}
\arguments{
\item{Metric}{A matrix supplied by user for which to calculate outliers within}

\item{alpha}{Alpha value for significance}
}
\value{
Outliers for all samples tested
}
\description{
calcOutliers - A function for calculating outliers for checkLM
}
