% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLM.R
\name{checkLM}
\alias{checkLM}
\title{check_LM - A function for determining whether an injection is an outlier}
\usage{
checkLM(
  filePath,
  dbName = "NameOfDB.db",
  instrument = "QTOF",
  projName = "",
  sampleMatrix = "",
  dPPM = 10,
  rtWin = 30,
  alpha = 0.01,
  noCheck = c("blank", "cond", "CP"),
  cwp = NULL,
  Config = Config,
  slackOn = F,
  batch = F
)
}
\arguments{
\item{filePath}{A filepath to the file to check}

\item{dbName}{Name of the DB to submit to}

\item{instrument}{Name of the type of instrument used for analysis (default is "QTOF)}

\item{projName}{If the user has supplied a projName it will be submitted to the DB}

\item{sampleMatrix}{If the user has supplied a sample matrix it will be submitted to the DB}

\item{dPPM}{The dPPM window for comparing peaks in sample against landmarks in DB}

\item{rtWin}{The RT window for comparing peaks in sample against landmarks in DB}

\item{alpha}{Specifies significance wanted for outlier detection}

\item{noCheck}{Part of sample names not to be compared to landmark database}

\item{cwp}{Option to specify custom centwave-parameters (see XCMS package). Otherwise uses default (View(check_LM))}

\item{Config}{Config object loaded using function readConfigFile}

\item{slackOn}{Boolean stating whether bad samples should be reported to slack or not}
}
\value{
Nothing
}
\description{
check_LM - A function for determining whether an injection is an outlier
}
