% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLMBatch.R
\name{checkLMBatch}
\alias{checkLMBatch}
\title{checkLMBatch - A function for performing checkLM on a batch of samples}
\usage{
checkLMBatch(fileDir, projName, cwp = NULL, Config, sumFiles, progressBatchRun)
}
\arguments{
\item{fileDir}{A filepath to the directory where the .mzML files which are to be checkLM analyzed are located}

\item{projName}{If the user has supplied a projName it will be submitted to the DB}

\item{cwp}{Option to specify custom centwave-parameters (see XCMS package). Otherwise uses default (View(check_LM))}

\item{Config}{Config object from readConfigFile function}

\item{sumFiles}{Tracking progress of checkLMBatch}

\item{progressBatchRun}{The dPPM window for comparing peaks in sample against landmarks in DB}
}
\value{
Nothing
}
\description{
checkLMBatch - A function for performing checkLM on a batch of samples
}
