% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLMBatchFolders.R
\name{checkLMBatchFolders}
\alias{checkLMBatchFolders}
\title{checkLMBatchFolders - A function for performing checkLM on a batch of samples}
\usage{
checkLMBatchFolders(dir, Config)
}
\arguments{
\item{dir}{One or several folders with mzML files to perform checkLMBatch on}

\item{Config}{A path to the config-file which will be used for the checkLMBatch procedures}
}
\value{
Nothing
}
\description{
checkLMBatchFolders - A function for performing checkLM on a batch of samples
}
