% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVToLM.R
\name{CSVtoLM}
\alias{CSVtoLM}
\title{CSVtoLM - A function for taking a .csv file and formatting it for submission to LM DB}
\usage{
CSVtoLM(filePath = "")
}
\arguments{
\item{filePath}{The path to the csv-file, based on the LMTemplate.csv file, with the LMs to be uploaded}
}
\value{
Returning data frame of LaMas
}
\description{
CSVtoLM - A function for taking a .csv file and formatting it for submission to LM DB
}
