% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropLandmarkViews.R
\name{dropLandmarkViews}
\alias{dropLandmarkViews}
\title{dropLandmarkViews - A function for removing landmark-views in the DB}
\usage{
dropLandmarkViews(dbName = "NameOfDB.db")
}
\arguments{
\item{dbName}{The name of the .db file in which views for landmarks will be setup}
}
\value{
Nothing
}
\description{
dropLandmarkViews - A function for removing landmark-views in the DB
}
