% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchIPOnLM.R
\name{fetchIPOnLM}
\alias{fetchIPOnLM}
\title{fetchIPOnLM - A function for fetching previous injections in the same seq.}
\usage{
fetchIPOnLM(dbName = "Test.db", toFetch)
}
\arguments{
\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function "buildDB()"}

\item{toFetch}{An object containing information on what to fetch}
}
\value{
Returning data frame of LaMa data
}
\description{
fetchIPOnLM - A function for fetching previous injections in the same seq.
}
