% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLMFeatureTable.R
\name{fetchLMFeatureTable}
\alias{fetchLMFeatureTable}
\title{fetchLMFeatureTable - A function for collecting LM feature data from a DB file}
\usage{
fetchLMFeatureTable(dbName = "NameOfDB.db", chromPol, projName, sampType)
}
\arguments{
\item{dbName}{Name of the DB to fetch from}

\item{chromPol}{The chromatography and ionization polarity of the files to fetch, expressed as :"RP", "RN", "HP" or "HN"}

\item{projName}{The projName of interest as it is stored in the DB file}

\item{sampType}{The type of injections of interest, "sQC" or "sample"}
}
\value{
Returning data frame of LaMa data from all samples fulfilling filters
}
\description{
fetchLMFeatureTable - A function for collecting LM feature data from a DB file
}
