% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLMRefDF.R
\name{fetchLMRefDF}
\alias{fetchLMRefDF}
\title{fetchLMRefDF - A function for fetching previous injections in the same seq.}
\usage{
fetchLMRefDF(dbName = "Test.db", toFetch)
}
\arguments{
\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function "buildDB()"}

\item{toFetch}{An object containing three variables: 'type' (containing 'sQC' or 'sample'), 'batchweek' (to determine which sequence the sample was injected in) and 'chromPol' (chromatography & pol; 'RP', 'RN', 'HP', 'HN')}
}
\value{
Returning table of LaMas for specified chromPol and sample type
}
\description{
fetchLMRefDF - A function for fetching previous injections in the same seq.
}
