% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLM.R
\name{find_LM}
\alias{find_LM}
\title{find_LM - A function to find landmarks in a XCMS object}
\usage{
find_LM(mode, XCMSObj, time_interval, mzdif, rtdif)
}
\arguments{
\item{mode}{For setting the name of the output}

\item{XCMSObj}{XCMS object after correspondence}

\item{time_interval}{Specifies how large retention time interval that the landmarks are selected in}

\item{mzdif}{The m/z difference allowed between features}

\item{rtdif}{The retention time difference allowed between features}
}
\value{
A list of the metadata of the landmarks.
}
\description{
find_LM - A function to find landmarks in a XCMS object
}
