% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{getTable - Function for extracting a peak-table from an XCMS object}
\usage{
getTable(XObj)
}
\arguments{
\item{XObj}{An XCMS object which has been processed to the "groupChromPeaks" step but before applying "fillChromPeaks"}
}
\value{
A peak table containing samples and features
}
\description{
getTable - Function for extracting a peak-table from an XCMS object
}
