% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOscore.R
\name{IPOscore}
\alias{IPOscore}
\title{IPOScore - Function for calculating isotope score for entire injection}
\usage{
IPOscore(xset, isotopeIdentification = c("IPO", "CAMERA"), ...)
}
\arguments{
\item{xset}{XCMS object}

\item{isotopeIdentification}{specify which identification algorithm to use, for more information, see IPO package}

\item{...}{???}
}
\value{
PPS value from IPO
}
\description{
IPOScore - Function for calculating isotope score for entire injection
}
