% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchQualiMon.R
\name{launchQualiMon}
\alias{launchQualiMon}
\title{launchApp - Function for running QualiMon}
\usage{
launchQualiMon()
}
\value{
shiny application object
}
\description{
launchApp - Function for running QualiMon
}
