% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeFindLM.R
\name{optimizeFindLM}
\alias{optimizeFindLM}
\title{optimizeFindLM - A function for optimization of finding new LaMas}
\usage{
optimizeFindLM(
  preFCPfilepath,
  split = "@",
  minrttocheck = 40,
  Prefilterintensity,
  mode,
  minLM = 100,
  maxLM = 200
)
}
\arguments{
\item{preFCPfilepath}{File path to a .rds file which contains an XCMS object which has been processed until, but not includining, gap-filling}

\item{split}{The character used in the XCMS object for splitting up mz & RT}

\item{minrttocheck}{Dead volume cut-off in beginning of chromatogram to disregard when looking for LaMas}

\item{Prefilterintensity}{Prefilter intensity settings used when peak picking with XCMS}

\item{mode}{???}
}
\value{
Returning 10 sets of optimized LaMas and the final parameters
}
\description{
optimizeFindLM - A function for optimization of finding new LaMas
}
