% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLMRT.R
\name{predictLMRT}
\alias{predictLMRT}
\title{predictLMRT - A function for building model of difference in RT between sample and DB}
\usage{
predictLMRT(LMRT, SampRT)
}
\arguments{
\item{LMRT}{RT of LaMas in DB}

\item{SampRT}{RT of LaMas in sample}
}
\value{
Returning predicted RTs of non-found LaMas
}
\description{
predictLMRT - A function for building model of difference in RT between sample and DB
}
