% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfigFile.R
\name{readConfigFile}
\alias{readConfigFile}
\title{readConfigFile - A function for reading config files}
\usage{
readConfigFile(configFilePath)
}
\arguments{
\item{configFilePath}{Filepath to config.txt file}
}
\value{
Config file object
}
\description{
readConfigFile - A function for reading config files
}
