% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalcNLMs.R
\name{recalcNLMs}
\alias{recalcNLMs}
\title{recalcNLMs - Function to recalculate nLMs from the information in the database. Only neccessary when code has been written wrongly in the first place.}
\usage{
recalcNLMs(dbName = "MedGICarbDB.db", chromPol, projName, sampType)
}
\arguments{
\item{dbName}{Name of the DB to fetch from}

\item{chromPol}{The chromatography and ionization polarity of the files to fetch, expressed as :"RP", "RN", "HP" or "HN"}

\item{projName}{The projName of interest as it is stored in the DB file}

\item{sampType}{The type of injections of interest, "sQC" or "sample"}
}
\value{
Nothing
}
\description{
recalcNLMs - Function to recalculate nLMs from the information in the database. Only neccessary when code has been written wrongly in the first place.
}
