% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyServer.R
\name{server}
\alias{server}
\title{Shiny app server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function
}
\author{
Anton Ribbenstedt
}
