% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLandmarkViews.R
\name{setupLandmarkViews}
\alias{setupLandmarkViews}
\title{setupLandmarkViews - A function for setting up landmark-views in the DB}
\usage{
setupLandmarkViews(dbName = "NameOfDB.db", path = getwd())
}
\arguments{
\item{dbName}{The name of the .db file in which views for landmarks will be setup}

\item{path}{Path where DB is located}
}
\value{
Nothing
}
\description{
setupLandmarkViews - A function for setting up landmark-views in the DB
}
