% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitBatchToDB.R
\name{submitBatchToDB}
\alias{submitBatchToDB}
\title{submitBatchToDB - A function for submitting injections to DB}
\usage{
submitBatchToDB(
  dir,
  projName,
  sampMatrix,
  dbName = "NameOfDB.db",
  instrument = "QTOF",
  cwp = NULL
)
}
\arguments{
\item{dir}{A diretion containing only .mzML files which are to be submitted to the DB}

\item{projName}{The project ID which will be assigned to all the entries into the DB}

\item{sampMatrix}{The sample matrix of all the samples that will be assigned to all the entries into the DB}

\item{dbName}{A list of names of the same length as the number of files in the direction 'dir'}

\item{instrument}{Description of instrument used to analyze the samples}

\item{cwp}{Option to specify custom centwave-parameters, see XCMS package.}
}
\value{
Nothing
}
\description{
submitBatchToDB - A function for submitting injections to DB
}
