% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitInjToDB.R
\name{submitInjToDB}
\alias{submitInjToDB}
\title{submitInjToDB - A function for submitting injections to DB}
\usage{
submitInjToDB(dbName = "NameOfDB.db", injToSub)
}
\arguments{
\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function "buildDB()"}

\item{injToSub}{A df with 9 columns: sampID, date, batchWeek, seqInj, chromPol, instrument, msMode, IPOscore and nLMs}
}
\value{
Returning status of whether injection already was in DB or not
}
\description{
submitInjToDB - A function for submitting injections to DB
}
