% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitLMPeaksToDB.R
\name{submitLMPeaksToDB}
\alias{submitLMPeaksToDB}
\title{submitLMPeaksToDB - A function for submitting injections to DB}
\usage{
submitLMPeaksToDB(dbName = "Test.db", injToSub)
}
\arguments{
\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function buildDB}

\item{injToSub}{A df with 9 columns: sampID, date, batchWeek, seqInj, chromPol, instrument, msMode, IPOscore and nLMs}
}
\value{
Nothing
}
\description{
submitLMPeaksToDB - A function for submitting injections to DB
}
