% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitLMQualityToDB.R
\name{submitLMQualityToDB}
\alias{submitLMQualityToDB}
\title{submitLMQualityToDB - A function for submitting iteration data for newly run samples}
\usage{
submitLMQualityToDB(testInj, dbName, qualityTable)
}
\arguments{
\item{testInj}{A df with 9 columns: sampID, date, batchWeek, seqInj, chromPol, instrument, msMode, IPOscore and nLMs}

\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function buildDB}

\item{qualityTable}{The table with the latest iteration of quality checks to be submitted to DB}
}
\value{
Nothing
}
\description{
submitLMQualityToDB - A function for submitting iteration data for newly run samples
}
