% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitSampToDB.R
\name{submitSampToDB}
\alias{submitSampToDB}
\title{submitSampToDB - A function for submitting samples to DB}
\usage{
submitSampToDB(dbName = "NameOfDB.db", injToSub)
}
\arguments{
\item{dbName}{Name / path to a SQLite of the same format as the example file. To make a new one see function "buildDB()"}

\item{injToSub}{A df with 9 columns: sampID, date, batchWeek, seqInj, chromPol, instrument, msMode, IPOscore and nLMs}
}
\value{
Nothing
}
\description{
submitSampToDB - A function for submitting samples to DB
}
