% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUi.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny app ui object}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
ui
}
\description{
Shiny app ui object
}
\author{
Anton Ribbenstedt
}
\keyword{datasets}
